cube: Try to bind to wl_compositor v4
VK_KHR_incremental_present requires the client to
bind to wl_compositor v4. This is because internally
some Vulkan Wayland drivers convert VkPresentRegionKHR
to wl_surface.damage_buffer()[1], which is a v4-only
request.

If the Wayland compositor advertises wl_compositor v4,
we can bind to it and make incremental present actually
work.

[1] https://gitlab.freedesktop.org/mesa/mesa/blob/master/src/vulkan/wsi/wsi_common_wayland.c#L862